pyproject.toml
[tool.poetry]
name = "subak-ckan"
version = "0.1.0"
description = "CKAN Helpers for Subak Data Catalog"
authors = ["Rabscuttler <laurence.watson@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
python-dotenv = "^0.19.0"
ipykernel = "^6.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
